+++ home-assistant-addon/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Bumped base image to debian:13.5
 - Fixed the issue that a closed release is still shown in the dropdown for releases in the board filter and the issuer modal
 - Allow single spaces in configurable column/status names
+- Added issue description confirmation also for new issues as for editing issues
 
 ## 1.3.0